refactor: Rename and repackage new screens
