+++ features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/abort/aborted/desktop/AbortedReturnToDesktopConstants.kt
@@ -1,0 +1,12 @@
+package uk.gov.onelogin.criorchestrator.features.handback.internal.abort.aborted.desktop
+
+import androidx.annotation.StringRes
+import uk.gov.onelogin.criorchestrator.features.handback.internal.R
+
+internal object AbortedReturnToDesktopConstants {
+    @StringRes
+    val titleId: Int = R.string.handback_confirmabortreturntodesktopweb_title
+
+    @StringRes
+    val buttonId: Int = R.string.handback_confirmaborttodesktopweb_button
+}
