+++ features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/abort/confirm/desktop/ConfirmAbortDesktopActions.kt
@@ -1,0 +1,5 @@
+package uk.gov.onelogin.criorchestrator.features.handback.internal.abort.confirm.desktop
+
+sealed class ConfirmAbortDesktopActions {
+    data object NavigateToReturnToDesktop : ConfirmAbortDesktopActions()
+}
