+++ features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/abort/confirm/desktop/ConfirmAbortDesktopConstants.kt
@@ -1,0 +1,12 @@
+package uk.gov.onelogin.criorchestrator.features.handback.internal.abort.confirm.desktop
+
+import androidx.annotation.StringRes
+import uk.gov.onelogin.criorchestrator.features.handback.internal.R
+
+internal object ConfirmAbortDesktopConstants {
+    @StringRes
+    val titleId: Int = R.string.handback_confirmabort_title
+
+    @StringRes
+    val buttonId: Int = R.string.handback_confirmaborttodesktopweb_button
+}
