+++ features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/abort/confirm/mobile/ConfirmAbortMobileAction.kt
@@ -1,0 +1,7 @@
+package uk.gov.onelogin.criorchestrator.features.handback.internal.abort.confirm.mobile
+
+sealed class ConfirmAbortMobileAction {
+    data class ContinueGovUk(
+        val redirectUri: String,
+    ) : ConfirmAbortMobileAction()
+}
