--- features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/confirmabort/confirmabortdesktop/ConfirmAbortDesktopActions.kt
@@ -1,5 +1,0 @@
-package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortdesktop
-
-sealed class ConfirmAbortDesktopActions {
-    data object NavigateToReturnToDesktop : ConfirmAbortDesktopActions()
-}
