--- features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/confirmabort/confirmabortdesktop/ConfirmAbortDesktopConstants.kt
@@ -1,12 +1,0 @@
-package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortdesktop
-
-import androidx.annotation.StringRes
-import uk.gov.onelogin.criorchestrator.features.handback.internal.R
-
-internal object ConfirmAbortDesktopConstants {
-    @StringRes
-    val titleId: Int = R.string.handback_confirmabort_title
-
-    @StringRes
-    val buttonId: Int = R.string.handback_confirmaborttodesktopweb_button
-}
