--- features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/confirmabort/confirmabortmobile/ConfirmAbortMobileAction.kt
@@ -1,7 +1,0 @@
-package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortmobile
-
-sealed class ConfirmAbortMobileAction {
-    data class ContinueGovUk(
-        val redirectUri: String,
-    ) : ConfirmAbortMobileAction()
-}
