--- features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/confirmabort/confirmabortreturntodesktop/ConfirmAbortReturnToDesktopConstants.kt
@@ -1,12 +1,0 @@
-package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortreturntodesktop
-
-import androidx.annotation.StringRes
-import uk.gov.onelogin.criorchestrator.features.handback.internal.R
-
-internal object ConfirmAbortReturnToDesktopConstants {
-    @StringRes
-    val titleId: Int = R.string.handback_confirmabortreturntodesktopweb_title
-
-    @StringRes
-    val buttonId: Int = R.string.handback_confirmaborttodesktopweb_button
-}
