--- features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/confirmabort/confirmabortreturntodesktop/ConfirmAbortReturnToDesktopViewModel.kt
@@ -1,16 +1,0 @@
-package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortreturntodesktop
-
-import androidx.lifecycle.ViewModel
-import uk.gov.onelogin.criorchestrator.features.handback.internal.analytics.HandbackAnalytics
-import uk.gov.onelogin.criorchestrator.features.handback.internal.analytics.HandbackScreenId
-
-class ConfirmAbortReturnToDesktopViewModel(
-    private val analytics: HandbackAnalytics,
-) : ViewModel() {
-    fun onScreenStart() {
-        analytics.trackScreen(
-            id = HandbackScreenId.ConfirmAbortReturnToDesktop,
-            title = ConfirmAbortReturnToDesktopConstants.titleId,
-        )
-    }
-}
