+++ features/handback/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/modal/AbortNavGraphProvider.kt
@@ -6,12 +6,12 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.squareup.anvil.annotations.ContributesMultibinding
-import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortdesktop.ConfirmAbortDesktopViewModelModule
-import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortdesktop.ConfirmAbortDesktopWebScreen
-import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortmobile.ConfirmAbortMobileScreen
-import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortmobile.ConfirmAbortMobileViewModelModule
-import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortreturntodesktop.ConfirmAbortReturnToDesktopViewModelModule
-import uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortreturntodesktop.ConfirmAbortReturnToDesktopWebScreen
+import uk.gov.onelogin.criorchestrator.features.handback.internal.abort.confirm.desktop.ConfirmAbortDesktopViewModelModule
+import uk.gov.onelogin.criorchestrator.features.handback.internal.abort.confirm.desktop.ConfirmAbortDesktopWebScreen
+import uk.gov.onelogin.criorchestrator.features.handback.internal.abort.confirm.mobile.ConfirmAbortMobileScreen
+import uk.gov.onelogin.criorchestrator.features.handback.internal.abort.confirm.mobile.ConfirmAbortMobileViewModelModule
+import uk.gov.onelogin.criorchestrator.features.handback.internal.abort.aborted.desktop.AbortedReturnToDesktopViewModelModule
+import uk.gov.onelogin.criorchestrator.features.handback.internal.abort.aborted.desktop.AbortedReturnToDesktopWebScreen
 import uk.gov.onelogin.criorchestrator.features.handback.internal.navigatetomobileweb.WebNavigator
 import uk.gov.onelogin.criorchestrator.features.handback.internalapi.nav.AbortDestinations
 import uk.gov.onelogin.criorchestrator.features.resume.internalapi.nav.ProveYourIdentityNavGraphProvider
@@ -27,7 +27,7 @@ class AbortNavGraphProvider
         private val confirmAbortToMobileWebViewModelFactory: ViewModelProvider.Factory,
         @Named(ConfirmAbortDesktopViewModelModule.FACTORY_NAME)
         private val confirmAbortToDesktopWebViewModelFactory: ViewModelProvider.Factory,
-        @Named(ConfirmAbortReturnToDesktopViewModelModule.FACTORY_NAME)
+        @Named(AbortedReturnToDesktopViewModelModule.FACTORY_NAME)
         private val confirmAbortReturnToDesktopWebViewModelFactory: ViewModelProvider.Factory,
         private val webNavigator: WebNavigator,
     ) : ProveYourIdentityNavGraphProvider {
@@ -51,7 +51,7 @@ class AbortNavGraphProvider
             }
 
             composable<AbortDestinations.ConfirmAbortReturnDesktop> {
-                ConfirmAbortReturnToDesktopWebScreen(
+                AbortedReturnToDesktopWebScreen(
                     viewModel = viewModel(factory = confirmAbortReturnToDesktopWebViewModelFactory),
                 )
             }
