+++ features/handback/internal/src/test/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/abort/aborted/desktop/AbortedReturnToDesktopScreenDisableBackButtonTest.kt
@@ -1,0 +1,24 @@
+package uk.gov.onelogin.criorchestrator.features.handback.internal.abort.aborted.desktop
+
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import org.junit.Before
+import org.junit.runner.RunWith
+import org.mockito.Mockito.mock
+import uk.gov.onelogin.criorchestrator.features.handback.internal.DisableBackButtonTest
+
+@RunWith(AndroidJUnit4::class)
+class AbortedReturnToDesktopScreenDisableBackButtonTest : DisableBackButtonTest() {
+    private val viewModel =
+        AbortedReturnToDesktopViewModel(
+            analytics = mock(),
+        )
+
+    @Before
+    fun setup() {
+        setContent {
+            AbortedReturnToDesktopWebScreen(
+                viewModel = viewModel,
+            )
+        }
+    }
+}
