--- features/handback/internal/src/test/kotlin/uk/gov/onelogin/criorchestrator/features/handback/internal/confirmabort/confirmabortdesktop/ConfirmAbortDesktopScreenAnalyticsTest.kt
@@ -1,80 +1,0 @@
-package uk.gov.onelogin.criorchestrator.features.handback.internal.confirmabort.confirmabortdesktop
-
-import android.content.Context
-import androidx.compose.ui.test.hasText
-import androidx.compose.ui.test.junit4.createComposeRule
-import androidx.compose.ui.test.performClick
-import androidx.test.core.app.ApplicationProvider
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import org.junit.Before
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.mockito.kotlin.mock
-import uk.gov.logging.api.v3dot1.logger.asLegacyEvent
-import uk.gov.logging.api.v3dot1.model.TrackEvent
-import uk.gov.logging.api.v3dot1.model.ViewEvent
-import uk.gov.onelogin.criorchestrator.features.handback.internal.analytics.HandbackAnalytics
-import uk.gov.onelogin.criorchestrator.features.handback.internal.analytics.HandbackScreenId
-import uk.gov.onelogin.criorchestrator.libraries.analytics.resources.AndroidResourceProvider
-import uk.gov.onelogin.criorchestrator.libraries.testing.ReportingAnalyticsLoggerRule
-import kotlin.test.assertContains
-
-@RunWith(AndroidJUnit4::class)
-class ConfirmAbortDesktopScreenAnalyticsTest {
-    @get:Rule
-    val composeTestRule = createComposeRule()
-
-    @get:Rule
-    val reportingAnalyticsLoggerRule = ReportingAnalyticsLoggerRule()
-    private val context: Context = ApplicationProvider.getApplicationContext()
-    private val analyticsLogger = reportingAnalyticsLoggerRule.analyticsLogger
-    private val continueButton = hasText(context.getString(ConfirmAbortDesktopConstants.buttonId))
-
-    private val analytics =
-        HandbackAnalytics(
-            resourceProvider = AndroidResourceProvider(context),
-            analyticsLogger = analyticsLogger,
-        )
-
-    private val viewmodel = ConfirmAbortDesktopViewModel(analytics)
-
-    @Before
-    fun setup() {
-        composeTestRule.setContent {
-            ConfirmAbortDesktopWebScreen(
-                viewModel = viewmodel,
-                navController = mock(),
-            )
-        }
-    }
-
-    @Test
-    fun `when screen starts, it tracks analytics`() {
-        val expectedEvent =
-            ViewEvent
-                .Screen(
-                    id = HandbackScreenId.ConfirmAbortDesktop.rawId,
-                    name = context.getString(ConfirmAbortDesktopConstants.titleId),
-                    params = HandbackAnalytics.requiredParameters,
-                ).asLegacyEvent()
-
-        assertContains(analyticsLogger.loggedEvents, expectedEvent)
-    }
-
-    @Test
-    fun `when continue button is clicked, it tracks analytics`() {
-        composeTestRule
-            .onNode(continueButton)
-            .performClick()
-
-        val expectedEvent =
-            TrackEvent
-                .Button(
-                    text = context.getString(ConfirmAbortDesktopConstants.buttonId),
-                    params = HandbackAnalytics.requiredParameters,
-                ).asLegacyEvent()
-
-        assertContains(analyticsLogger.loggedEvents, expectedEvent)
-    }
-}
